Optimization when not sliding
The current version of this mod doesn't perform sliding if the Crouching checks fail, but it still processes some of the sliding logic anyways. To account for this, the Crouching check has been moved to the start of the CheckSliding hook.

Also made all of the hook names more unique and localized some global calls